+++ NEWS.md
@@ -16,6 +16,8 @@ by using the reference semantic of the data.table package. If many columns are a
 - `p_attributes()` implemented for `remote_corpus` and `remote_partition`.
 - A new `regions()` method (for `corpus` class objects to start with) returns a `regions` class object with a regions matrix (slot `cpos`) with regions for an s-attribute (#176).
 - The `get_token_stream()`-method for `regions` and `matrix` objects will now accept a logical argument `split`. If `TRUE`, a list of character vectors is returned. The envisaged use case is a fast decoding of sentences (#176).
+- A `encoding()` method has been defined if argument `object` is missing. Calling `encoding()` will return the session character set. If it cannot be determined using `localeToCharset()`, a UTF-8 session charset will be assumed. Internally, `encoding()` replaces a direct call of `localeToCharset()` to avoid errors that have occurred on GitHub Actions with Ubuntu 20.04 (#188).
+- If the session character set cannot be guessed by `localeToCharset()` (`NA` return value), a startup message will issue a warning that 'UTF-8' is assumed (#188).
 
 ## Minor Improvements